Improve query by scapping special characters to replace and fix LongAdapter (mongo expects a String when using longs)

+++ jsongo/src/test/java/me/googas/jsongo/models/Note.java
@@ -10,7 +10,7 @@ public class Note {
   @NonNull
   @SerializedName("_id")
   @Getter
-  private final int id;
+  private final long id;
 
   @NonNull
   @SerializedName("user")
@@ -19,7 +19,7 @@ public class Note {
 
   @Getter @Setter private String value;
 
-  public Note(int id, @NonNull String userId, String value) {
+  public Note(long id, @NonNull String userId, String value) {
     this.id = id;
     this.userId = userId;
     this.value = value;
